pyproject.toml
[project]
name = "paperflux"
version = "0.1.0"
description = ""
authors = [
    {name = "kartikbhtt7",email = "kartikbhtt7@gmail.com"},
    {name = "Vector73",email = "v.shm.kunal@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio (>=5.16.0,<6.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "redis (>=5.2.1,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)",
    "markdown (>=3.7,<4.0)",
    "pymongo (>=4.11.1,<5.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "aiohttp (>=3.11.12,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"
black = "^25.1.0"
flake8 = "^7.1.1"

